Make pdbId mandatory, add fallback organism
