+++ app/assets/javascripts/3dbio_viewer/src/data/repositories/protvista/ebi-pdbe-api.ts
@@ -1,10 +1,15 @@
 import { Experiment } from "../../../domain/entities/Experiment";
 
 export type PdbExperiment = Record<string, ExperimentInfo[]>;
+export type PdbMolecules = Record<string, MoleculeInfo[]>;
 
 interface ExperimentInfo {
     resolution: number;
     experimental_method: string;
+}
+
+interface MoleculeInfo {
+    source: { organism_scientific_name: string }[];
 }
 
 export function getExperiment(pdbId: string, pdbExperiment: PdbExperiment): Experiment | undefined {
@@ -15,4 +20,8 @@ export function getExperiment(pdbId: string, pdbExperiment: PdbExperiment): Expe
         resolution: info.resolution,
         method: info.experimental_method,
     };
+}
+
+export function getFallbackOrganism(pdbId: string, pdbMolecules: PdbMolecules) {
+    return pdbMolecules[pdbId]?.[0]?.source?.[0]?.organism_scientific_name;
 }
