+++ app/assets/javascripts/3dbio_viewer/src/domain/entities/Pdb.ts
@@ -33,6 +33,7 @@ export interface Pdb {
     customAnnotations: Maybe<Annotations>;
     ligands: Maybe<PdbLigand[]>;
     publications: PdbPublication[];
+    organism: Maybe<string>;
 }
 
 export interface PdbLigand {
