+++ app/assets/javascripts/3dbio_viewer/src/domain/repositories/PdbRepository.ts
@@ -10,6 +10,6 @@ export interface PdbRepository {
 
 export interface PdbOptions {
     proteinId: Maybe<ProteinId>;
-    pdbId: Maybe<PdbId>;
+    pdbId: PdbId;
     chainId: ChainId;
 }
